Resolution of Issue #2365
Discussion on 8/19/22 to resolve 2 technical issues present in the slicing of practitioner.qualification - a value set had incorrectly been copied into both slices (though the intent was to maintain separation based on code system), and a value set was used in both rather than a code system in the slices. The original code systems were input and an STU note was added to summarize discussion.